refactor: replace shared model and repository with base model and repository, update schemas and controllers for consistency
