+++ projeto_aplicado/resources/__init__.py
@@ -1,8 +1,8 @@
 from sqlmodel import SQLModel
 
+from .base.model import BaseModel
 from .order.model import Order, OrderItem
 from .product.model import Product
-from .shared.model import BaseModel
 from .user.model import User
 
 __all__ = [
@@ -21,9 +21,9 @@ def get_metadata():
     This function ensures all models are imported and registered with SQLModel's metadata.
     """  # noqa: E501
     # Import all models to ensure they are registered
+    from .base.model import BaseModel  # noqa: F401, PLC0415
     from .order.model import Order, OrderItem  # noqa: F401, PLC0415
     from .product.model import Product  # noqa: F401, PLC0415
-    from .shared.model import BaseModel  # noqa: F401, PLC0415
     from .user.model import User  # noqa: F401, PLC0415
 
     return SQLModel.metadata
