+++ projeto_aplicado/resources/user/controller.py
@@ -4,7 +4,7 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException
 
 from projeto_aplicado.auth.security import get_current_user
-from projeto_aplicado.resources.shared.schemas import Pagination
+from projeto_aplicado.resources.base.schemas import Pagination
 from projeto_aplicado.resources.user.model import User, UserRole
 from projeto_aplicado.resources.user.repository import (
     UserRepository,
